Initial pass on a pdf-archive system.

+++ cover/models.py
@@ -5,7 +5,7 @@ from django.db import models
 class Author(models.Model):
     first_name = models.CharField(max_length=30)
     last_name  = models.CharField(max_length=40)
-    year = models.IntegerField()
+    year = models.PositiveSmallIntegerField()
 
     def __str__(self):
         return '%s %s' % (self.first_name, self.last_name)
